+++ src/Exception/ApiException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\os2forms_organisation\Exception;
+
+/**
+ * API exception.
+ */
+class ApiException extends Exception {
+
+}